server/myproject/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.9.0
